gardenbook_chat/requirements.txt
mcp-python
langchain-core
langchain-anthropic
langchain-ollama
langgraph
langchain-mcp-adapters
python-dotenv
pymongo